Refactor coordinates.
